Trait wright::errors::WrightError

pub trait WrightError {
    fn new(info: String, level: WrightErrorLevels) -> Self;

    fn get_info(&self) -> String;

    fn get_type(&self) -> String;

    fn get_level(&self) -> WrightErrorLevels;

    fn panic(&self) -> i32 { ... }
}

Trait for Wright's compiler errors. Each error should have its own implementation of this.

Required Methods

fn new(info: String, level: WrightErrorLevels) -> Self

The constructor method.

fn get_info(&self) -> String

Returns information about the error as a String. Usually a formatted output statement referencing the line, and sometimes location, where the error was raised.

fn get_type(&self) -> String

Return the type of error, usually a short error name like "IOError" or "ParserError"

fn get_level(&self) -> WrightErrorLevels

Returns the error level as an instance of WrightErrorLevels

Provided Methods

fn panic(&self) -> i32

Executes the error, terminating the program and returning a failed exit code (1).